test: add test cases for cupy inputs
This commit just adds some test cases to cover cupy inputs. The only
exception is the function test_output_match_cp(). It needs to be
converted to numpy because pytest.approx only supports numpy. As it is
just a test case, there is no problem to get data from GPU to test.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>